# Conflicts:
#	src/main/java/models/Commands/GameMenuCommands.java
#	src/test/java/LoginTest.java

+++ src/main/java/controllers/GameMenuController.java
@@ -186,9 +186,9 @@ public class GameMenuController implements Controller {
         double energyCost = (double) distance / 20 + (double) changedDir / 2;
         if (distance == 0)
             return new WalkProposal(false, "you can't reach " + goal.getPosition(), 0, x, y);
+        if (!player.doesOwnTile(goal))
+            return new WalkProposal(false, "you can't go in " + goal.getOwner().getAccount().getNickname() + "'s farm", 0, x, y);
         return new WalkProposal(true, "OK", energyCost, x, y);
-
-
     }
 
     private static int getChangedDir(List<Tile> path) {
@@ -209,7 +209,7 @@ public class GameMenuController implements Controller {
     public Result executeWalk(WalkProposal p) {
         Player player = App.getActiveGame().getCurrentPlayer();
         if (!p.isAllowed()) {
-            return new Result(false, "No walk was proposed");
+            return new Result(false, p.message());
         }
         player.setPosition(new Position(p.x(), p.y()));
         player.reduceEnergy(p.energyCost());
@@ -253,7 +253,8 @@ public class GameMenuController implements Controller {
 
                 if (newCol >= 0 && newCol < cols && newRow >= 0 && newRow < rows) {
                     Tile neighbor = tiles[newRow][newCol];
-                    if (neighbor.getContent() == null || neighbor.getContent().getComponent(Placeable.class).isWalkable())
+                    if ((neighbor.getContent() == null || neighbor.getContent().getComponent(Placeable.class).isWalkable()) &&
+                            (neighbor.getType() != null && neighbor.getType().isWalkable))
                         if (!cameFrom.containsKey(neighbor)) {
                             queue.add(neighbor);
                             cameFrom.put(neighbor, current);
@@ -274,8 +275,6 @@ public class GameMenuController implements Controller {
         }
         Collections.reverse(path);
         return path;
-
-
     }
 
     public Result helpReadingMap() {
@@ -492,11 +491,15 @@ public class GameMenuController implements Controller {
         if (position == null) {
             return new Result(false, "type a valid direction");
         }
-        if (!player.getComponent(Inventory.class).doesHaveItem(seedDetails.getEntityName())) {
+        if (!player.getComponent(Inventory.class).doesHaveItem(seedString)) {
             return new Result(false, "you don't have that seed");
         }
 
-        Entity seed = player.getComponent(Inventory.class).takeFromInventory(seedDetails, 1);
+
+        if(!player.getComponent(Inventory.class).doesHaveItem(seedString)){
+            return new Result(false, "you don't have that seed");
+        }
+        Entity seed = App.entityRegistry.getEntityDetails(seedString);
 
         Tile tile = game.getActiveMap().getTileByPosition(position);
 
@@ -508,17 +511,25 @@ public class GameMenuController implements Controller {
             return new Result(false, "tile isn't empty");
         }
         Entity building = tile.getMap().getBuilding();
-        if(building != null && StringUtils.isNamesEqual(building.getEntityName(), "greenhouse")){
-            //TODO greenhouse logic (planting in all seasons, cant plant trees, ...)
+
+        boolean canPlant = true;
+        Entity plant = seed.getComponent(SeedComponent.class).getGrowingPlant();
+
+        if(!plant.getComponent(Growable.class).getGrowingSeasons().contains(game.getDate().getSeason())){
+            canPlant = false;
         }
 
-//        Entity plant = App.entityRegistry.makeEntity(seed.getComponent(SeedComponent.class).getGrowingPlant());
+        if(building != null && StringUtils.isNamesEqual(building.getEntityName(), "greenhouse")){
+            canPlant = true;
+        }
 
-        Entity plant = seed.getComponent(SeedComponent.class).getGrowingPlant();
+        if(!canPlant){
+            return new Result(false, "You can't plant a " + plant.getEntityName() + " in this season");
+        }
+
+        seed = player.getComponent(Inventory.class).takeFromInventory(seedString, 1);
         tile.setType(TileType.PLANTED_GROUND);
         EntityPlacementSystem.placeOnTile(plant, tile);
-
-
 
         return new Result(true, "planted succusfully");
     }
@@ -1150,7 +1161,7 @@ public class GameMenuController implements Controller {
 
         }
 
-        skill.addExperience(5);
+        skill.addExperience(10);
         return new Result(true, message.toString());
     }
 
@@ -1671,23 +1682,19 @@ public class GameMenuController implements Controller {
             case 'a':
             case 'A':
                 p = this.proposeWalk(player.getPosition().getCol() - 1, player.getPosition().getRow());
-                executeWalk(p);
-                break;
+                return executeWalk(p);
             case 's':
             case 'S':
                 p = this.proposeWalk(player.getPosition().getCol(), player.getPosition().getRow() + 1);
-                executeWalk(p);
-                break;
+                return executeWalk(p);
             case 'w':
             case 'W':
                 p = this.proposeWalk(player.getPosition().getCol(), player.getPosition().getRow() - 1);
-                executeWalk(p);
-                break;
+                return executeWalk(p);
             case 'd':
             case 'D':
                 p = this.proposeWalk(player.getPosition().getCol() + 1, player.getPosition().getRow());
-                executeWalk(p);
-                break;
+                return executeWalk(p);
 
 
             case 'x':
@@ -1963,7 +1970,6 @@ public class GameMenuController implements Controller {
     }
     public Result buildGreenhouse(){
         Player player = App.getActiveGame().getCurrentPlayer();
-        player.getOwnedPlantedTiles();
 
         int wood = player.getComponent(Inventory.class).getItemCount("Wood");
         double money = player.getWallet().getBalance();
@@ -1981,6 +1987,9 @@ public class GameMenuController implements Controller {
 
         EntityPlacementSystem.clearArea(position.getCol(), position.getRow(), greenhouse.getComponent(Placeable.class));
         EntityPlacementSystem.placeEntity(greenhouse, position);
+
+        player.getWallet().reduceBalance(1000);
+        player.getComponent(Inventory.class).takeFromInventory("Wood", 500);
 
         return new Result(true, "greenhouse built");
     }
@@ -2006,6 +2015,7 @@ public class GameMenuController implements Controller {
     public Result trashItem(String name, int amount){
         Player player = App.getActiveGame().getCurrentPlayer();
         Entity entity = player.getComponent(Inventory.class).takeFromInventory(name, player.getComponent(Inventory.class).getItemCount(name));
+        if(entity == null) return new Result(false, "");
         entity.delete();
         //TODO
         return new Result(false, "");
